requirement.txt
Django==5.1.6
djangorestframework==3.15.2
pandas==2.2.3
redis==5.2.1
pytest-django==4.10.0
celery==5.4.0
coverage==7.6.12